Added functionality to join via ID, added bidirectional relationships for Users and Game, tried to get Websockets to work (not working yet)
Users can now join game via ID, and are added to a list saved in the game entity. The user and game entities now have bidirectional relationships (currently a bit vulnerable to recursion). A user can only be assigned to a single game, A game can have many users. A user can only host one game and a game can only have one host. Additionally, I wrote the start of what's supposed to be a SPRING websocket to have a dynamically changing user list in the waiting room in the frontend.